requirements.txt
requests==2.32.5
pysocks==1.7.1
configparser==7.2.0
argcomplete==3.6.2
colorama==0.4.6
rich==13.9.4